tests/wpt/tests/html/semantics/interestfor/interestfor-user-select.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui/#propdef-user-select">
<link rel="help" href="https://open-ui.org/components/interest-invokers.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src='/resources/testdriver-vendor.js'></script>

<button interestfor=foo>ABCDEFGHI<span id=inside>JKLMNOPQ</span>RSTUVWXYZ</button>
<div>Something after the button</div>
<div id=after>Something else</div>

<script>
promise_test(async function() {
  assert_equals(window.getSelection().toString(), "", "Nothing should start out selected");

  await new test_driver.Actions()
    .pointerMove(1, 1, {origin: after})
    .pointerDown({ button: 0 })
    .pointerMove(1, 1, {origin: inside})
    .pointerUp({ button: 0 })
    .send();
  const selection = window.getSelection().toString();
  assert_not_equals(selection, "", "Something should be selected");
  assert_true(selection.includes("Something after"),'The selection should include the outside text');
  assert_false(selection.includes("RSTUVWXYZ"),'The selection should not include the button text');
}, "Buttons with `interestfor` should not be user-selectable");
</script>
